requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
black==24.1.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
cssutils==2.11.1
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-cli==0.0.5
fastapi-sessions==0.3.2
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.65.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
kombu==5.4.2
llama-cloud==0.0.12
llama-index==0.10.62
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.62
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.29
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.32
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-pinecone==0.1.9
llama-parse==0.4.9
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
memory-profiler==0.61.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.54.4
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pgvector==0.3.6
pillow==11.0.0
pinecone==5.4.2
pinecone-plugin-assistant==1.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
protoc-gen-openapiv2==0.0.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.8.0
tomli==2.1.0
transformers==4.48.0
tqdm==4.67.0
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
wrapt==1.16.0
yarl==1.17.2
